pyproject.toml
[tool.poetry]
name = "py-acgimg-bot"
version = "0.1.0"
description = ""
authors = ["Expliyh <expliyh@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.115.8"
uvicorn = "0.34.0"
python-telegram-bot = {extras = ["all"], version = "21.10"}
sqlalchemy = {extras = ["asyncio"], version = "2.0.38"}
asyncmy = "0.2.10"
mariadb = "1.1.11"
python-dotenv = "^1.0.0"
singleton-class-decorator = "^1.0.1"
aiofiles = "24.1.0"
aiohttp = "3.11.12"
aiorwlock = "1.5.0"
pillow = "11.1.0"
pixivpy3 = "3.7.5"
b2sdk = "2.8.0"
redis = "5.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
